Deprecate the `host` field, and introduce new `client_ip` field instead:
This is to avoid a naming conflict with the ELK stack, specifically filebeat.
From version 6.3.x onwards, filebeat enriches the events it reads with
information about the host that filebeat is running on, and will write
that metadata to the `host` field of the event (hardcoded).

This will overwrite any actual payload data in the event's `host` field.

See https://www.elastic.co/guide/en/beats/libbeat/6.3/breaking-changes-6.3.html#breaking-changes-mapping-conflict
for some details on this change.

There are some ways to work around this, and change the filebeat configuration
so that the event's own `host` field gets preserved, but with the inclusion
of the `host` field in ECS (Elastic Common Schema) it's quite likely that
more conflicts around this field might arise in the future, especially if it's
being used with a different type than defined in ECS (object vs. string):
https://github.com/elastic/ecs/#host

With filebeat officially adding this to the list of exported fields, it's
also possible that components down the line (i.e., logstash) might be changed
in a way where they rely on this field containing the actual host metadata:
https://www.elastic.co/guide/en/beats/filebeat/current/exported-fields-host-processor.html

Therefore we deprecate the `host` field in ftw.structlog, and introduce a new
field `client_ip` with the same semantics that should be used instead.