Environment variable for muting eventlog setup errors.
When running tests of other projects there is usually no eventlog
configured. ftw.structlog then spams the test output.

For muting the setup errors in this situation an environment variable
is introduced. The variable is only meant to be used in testing since
we want to see the errors in production.