Fluentd logging: Use K8s namespace in tag:
This uses the K8s namespace to build the tag that is used for Fluentd logging.

Since this tag is used by the Fluentd file output as the file name, this makes
sure the resulting file name is dynamic and logs for deployments in different
K8s namespaces are logged to separate files.

In order for this to work, the K8s namespace must be exposed to the container
in the env var KUBERNETES_NAMESPACE.